Convert ISIC Codes

Convert International Standard Industrial Classification (ISIC) codes to aggregate economic activity categories.

Usage

wid_convert_isic(codes, level = 4L, full_label = FALSE)

Arguments

codes

A character vector of original ISIC codes, which can be at the section, division, group, class, or 4-digit level.

level

Level of dis-aggregation (determining number of categories):

1	Section-level (default; 21 categories).

2	Lowest activity-level (6 categories).

3	4 categories.

4	3 categories.

5	Agricultural or not (2 categories).

full_label

Logical; if TRUE, will return full text labels rather than short codes.

Value

A character vector containing short code or labels.

Examples

(codes <- LETTERS[1:21])
#>  [1] "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "N" "O" "P" "Q" "R" "S"
#> [20] "T" "U"
wid_convert_isic(codes)
#>  [1] "A" "I" "I" "I" "I" "I" "S" "S" "S" "S" "S" "S" "S" "S" "S" "S" "S" "S" "S"
#> [20] "S" "S"
wid_convert_isic(codes, 1, full_label = TRUE)
#>  [1] "Agriculture"      "Mining"           "Manufacturing"    "Electrical"      
#>  [5] "Water"            "Construction"     "Trade"            "Transportation"  
#>  [9] "Service"          "Information"      "Financial"        "Property"        
#> [13] "Professional"     "Administrative"   "Public"           "Education"       
#> [17] "Health"           "Arts"             "Other"            "Households"      
#> [21] "Extraterritorial"
wid_convert_isic(codes, 2, full_label = TRUE)
#>  [1] "Agriculture"              "Industry - Resources"    
#>  [3] "Industry - Manufacturing" "Industry - Resources"    
#>  [5] "Industry - Resources"     "Industry - Construction" 
#>  [7] "Services - Market"        "Services - Market"       
#>  [9] "Services - Market"        "Services - Market"       
#> [11] "Services - Market"        "Services - Market"       
#> [13] "Services - Market"        "Services - Market"       
#> [15] "Services - Non-Market"    "Services - Non-Market"   
#> [17] "Services - Non-Market"    "Services - Non-Market"   
#> [19] "Services - Non-Market"    "Services - Non-Market"   
#> [21] "Services - Non-Market"